chore: took in account feedback
